requirements.txt
setuptools>=36,<46
cython>=0.2
pyzmq>=16.0.2,<17.0
csirtg_indicator>=1.0.1,<2.0
cifsdk>=3.0.0rc4,<4.0
Flask-Limiter>=0.9.4,<1.0
limits>=1.1.1,<1.2
maxminddb>=1.2.0,<1.3
geoip2>=2.8.0,<2.9
pygeoip==0.3.2
dnspython>=1.15.0,<1.16
Flask>=1.0
flask-cors>=3.0,<4.0
PyYAML>=4.2b1
SQLAlchemy>=1.0.14,<1.1
elasticsearch>=5.3,<5.5
elasticsearch-dsl>=5.3,<5.5
html5lib==1.0b8 # bug in csirtg-smrt upstream
msgpack-python>=0.4.8,<0.5.0
apwgsdk==0.0.0a6
csirtg_smrt>=1.0,<2.0
csirtg_dnsdb==0.0.0a4
tornado>=5.1.0
faker==0.7.10
Flask-Bootstrap==3.3.6.0
gevent~=21.1.2
gunicorn==19.7.1
urllib3<1.25,>=1.21.1
requests==2.22.0
ujson<=4.0.2
zipp==0.6.0
